Urls refactored #141.
