+++ AlertaDengue/dados/templates/alerta_mrj.html
@@ -58,7 +58,7 @@
     <script src="/static/js/QuadTree.js"></script>
     <script src='/static/exp_PontosUnidadesBasicasSaudeMRJ2014wgs84.js'></script>
     <script src="/static/js/Autolinker.min.js"></script>
-    <script src="/static/js/jquery.sparkline.min.js"></script>
+    <script src="/static/libs/jquery/jquery.sparkline.min.js"></script>
     <script src="/static/js/highstock.js"></script>
     <script src="/static/js/modules/exporting.js"></script>
     <style>
